pyproject.toml
[project]
name = "arbbuilder"
version = "0.1.0"
description = "AI-powered development assistant for the Arbitrum ecosystem"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "ARBuilder Team"},
]
keywords = ["arbitrum", "stylus", "rust", "smart-contracts", "rag", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Code Generators",
]

dependencies = [
    "crawl4ai>=0.7.0",
    "chromadb>=1.0.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.9.0",
    "beautifulsoup4>=4.12.0",
    "markdownify>=0.12.0",
    "tiktoken>=0.5.0",
    "pydantic>=2.0.0",
    "rich>=13.0.0",
    "httpx>=0.27.0",
    "tenacity>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
arbbuilder-scrape = "scraper.run:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["scraper*", "src*"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W"]
